♻️ refactor(wallets): remove cyclic-import disables and reorganize imports
