+++ services/web/server/src/simcore_service_webserver/payments/_onetime_api.py
@@ -1,4 +1,3 @@
-# pylint: disable=cyclic-import  # Deferred imports at function level only; safe
 import logging
 from decimal import Decimal
 from typing import Any
