+++ services/web/server/src/simcore_service_webserver/users/_users_service.py
@@ -1,4 +1,3 @@
-# pylint: disable=cyclic-import  # Deferred imports at function level only; safe
 import logging
 from typing import Any
