+++ services/web/server/src/simcore_service_webserver/wallets/_handlers.py
@@ -15,10 +15,9 @@ from models_library.api_schemas_webserver.wallets import (
     WalletGet,
     WalletGetWithAvailableCredits,
 )
-from models_library.rest_base import RequestParameters, StrictRequestParameters
+from models_library.rest_base import RequestParameters
 from models_library.rest_error import ErrorGet
 from models_library.users import UserID
-from models_library.wallets import WalletID
 from pydantic import Field
 from servicelib.aiohttp import status
 from servicelib.aiohttp.request_keys import RQT_USERID_KEY
@@ -36,16 +35,26 @@ from ..exception_handling import (
     create_error_response,
 )
 from ..login.decorators import login_required
-from ..payments import payments_service
+from ..payments.errors import (
+    InvalidPaymentMethodError,
+    PaymentCompletedError,
+    PaymentMethodAlreadyAckedError,
+    PaymentMethodNotFoundError,
+    PaymentMethodUniqueViolationError,
+    PaymentNotFoundError,
+    PaymentServiceUnavailableError,
+    PaymentUniqueViolationError,
+)
 from ..products.errors import BelowMinimumPaymentError, ProductPriceNotDefinedError
 from ..security.decorators import permission_required
-from ..users import users_service
+from ..users.exceptions import BillingDetailsNotFoundError, UserDefaultWalletNotFoundError
 from ..utils_aiohttp import envelope_json_response
 from . import _api
 from ._constants import (
     MSG_BILLING_DETAILS_NOT_DEFINED_ERROR,
     MSG_PRICE_NOT_DEFINED_ERROR,
 )
+from ._schemas import WalletsPathParams
 from .errors import (
     WalletAccessForbiddenError,
     WalletNotEnoughCreditsError,
@@ -94,9 +103,9 @@ def handle_wallets_exceptions(handler: Handler):  # noqa: C901
 
         except (
             WalletNotFoundError,
-            payments_service.PaymentNotFoundError,
-            payments_service.PaymentMethodNotFoundError,
-            users_service.UserDefaultWalletNotFoundError,
+            PaymentNotFoundError,
+            PaymentMethodNotFoundError,
+            UserDefaultWalletNotFoundError,
         ) as exc:
             raise web.HTTPNotFound(text=f"{exc}") from exc
 
@@ -109,15 +118,15 @@ def handle_wallets_exceptions(handler: Handler):  # noqa: C901
             )
 
         except (
-            payments_service.PaymentUniqueViolationError,
-            payments_service.PaymentCompletedError,
-            payments_service.PaymentMethodAlreadyAckedError,
-            payments_service.PaymentMethodUniqueViolationError,
-            payments_service.InvalidPaymentMethodError,
+            PaymentUniqueViolationError,
+            PaymentCompletedError,
+            PaymentMethodAlreadyAckedError,
+            PaymentMethodUniqueViolationError,
+            InvalidPaymentMethodError,
         ) as exc:
             raise web.HTTPConflict(text=f"{exc}") from exc
 
-        except payments_service.PaymentServiceUnavailableError as exc:
+        except PaymentServiceUnavailableError as exc:
             return _create_error_response_with_support_id_and_logging(
                 request,
                 exc,
@@ -137,7 +146,7 @@ def handle_wallets_exceptions(handler: Handler):  # noqa: C901
         except WalletNotEnoughCreditsError as exc:
             raise web.HTTPPaymentRequired(text=f"{exc}") from exc
 
-        except users_service.BillingDetailsNotFoundError as exc:
+        except BillingDetailsNotFoundError as exc:
             raise web.HTTPServiceUnavailable(text=MSG_BILLING_DETAILS_NOT_DEFINED_ERROR) from exc
 
     return wrapper
@@ -152,10 +161,6 @@ routes = web.RouteTableDef()
 class WalletsRequestContext(RequestParameters):
     user_id: UserID = Field(..., alias=RQT_USERID_KEY)  # type: ignore[literal-required]
     product_name: str = Field(..., alias=RQ_PRODUCT_KEY)  # type: ignore[literal-required]
-
-
-class WalletsPathParams(StrictRequestParameters):
-    wallet_id: WalletID
 
 
 @routes.post(f"/{VTAG}/wallets", name="create_wallet")
