+++ services/web/server/src/simcore_service_webserver/wallets/models.py
@@ -2,17 +2,9 @@
 
 from ._groups_api import WalletGroupGet
 from ._groups_db import WalletGroupGetDB
-from ._groups_handlers import (
-    _WalletsGroupsBodyParams,
-    _WalletsGroupsPathParams,
-)
-from ._handlers import WalletsPathParams
 
 __all__: tuple[str, ...] = (
     # models
     "WalletGroupGet",
     "WalletGroupGetDB",
-    "WalletsPathParams",
-    "_WalletsGroupsBodyParams",
-    "_WalletsGroupsPathParams",
 )
